school_erp/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
python-dotenv==1.0.1
passlib[argon2]==1.7.4
email-validator==2.2.0
APScheduler==3.10.4
sqlitecloud==0.0.84
sqlalchemy-sqlitecloud==0.1.2
openai==1.68.2
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
reportlab==4.2.2
pytest==8.3.2
pytest-flask==1.3.0
gunicorn==22.0.0
